Payload

+++ nfqueue.go
@@ -131,7 +131,7 @@ func go_nfq_callback(id uint32, hwproto uint16, hook uint8, mark *uint32,
 	version, protocol, tos, ttl uint8, saddr, daddr unsafe.Pointer,
 	sport, dport, checksum uint16, payload unsafe.Pointer, payload_len int, nfqptr unsafe.Pointer) (v uint32) {
 
-	data := C.GoBytes(payload, C.int(payload_len))
+	data := C.GoBytes(payload, payload_len)
 
 	var (
 		nfq   = (*nfQueue)(nfqptr)
